REFPLTV-1627: RPI 4 64 Bit Build Broken
Reason for change: Fixed build issue for rialto for RPI 4 64 bit hybrid build.

Test Procedure: RPI 4 64 Bit Build for rialto should be successful.

Risks: Low

+++ source/GStreamerWebAudioPlayerClient.cpp
@@ -253,7 +253,7 @@ void GStreamerWebAudioPlayerClient::pushSamples()
     {
         return;
     }
-    uint32_t availableFrames = 0u;
+    uintptr_t availableFrames = 0u;
     if (mClientBackend->getBufferAvailable(availableFrames))
     {
         auto dataToPush = std::min(availableFrames * m_frameSize, mSampleDataBuffer.size());
