+++ source/WebAudioClientBackendInterface.h
@@ -32,7 +32,7 @@ public:
     virtual bool play() = 0;
     virtual bool pause() = 0;
     virtual bool setEos() = 0;
-    virtual bool getBufferAvailable(uint32_t &availableFrames) = 0;
+    virtual bool getBufferAvailable(uintptr_t &availableFrames) = 0;
     virtual bool getBufferDelay(uint32_t &delayFrames) = 0;
     virtual bool writeBuffer(const uint32_t numberOfFrames, void *data) = 0;
     virtual bool getDeviceInfo(uint32_t &preferredFrames, uint32_t &maximumFrames, bool &supportDeferredPlay) = 0;
